+++ packages/social-controllers/src/SocialController-method-action-types.ts
@@ -57,10 +57,28 @@ export type SocialControllerUpdateFollowingAction = {
 };
 
 /**
+ * Opts the current user out of the PnL leaderboard.
+ */
+export type SocialControllerOptOutOfLeaderboardAction = {
+  type: `SocialController:optOutOfLeaderboard`;
+  handler: SocialController['optOutOfLeaderboard'];
+};
+
+/**
+ * Opts the current user back into the PnL leaderboard.
+ */
+export type SocialControllerOptInToLeaderboardAction = {
+  type: `SocialController:optInToLeaderboard`;
+  handler: SocialController['optInToLeaderboard'];
+};
+
+/**
  * Union of all SocialController action types.
  */
 export type SocialControllerMethodActions =
   | SocialControllerUpdateLeaderboardAction
   | SocialControllerFollowTraderAction
   | SocialControllerUnfollowTraderAction
-  | SocialControllerUpdateFollowingAction;
+  | SocialControllerUpdateFollowingAction
+  | SocialControllerOptOutOfLeaderboardAction
+  | SocialControllerOptInToLeaderboardAction;